+++ cookbooks/aws-parallelcluster-platform/spec/unit/recipes/disable_services_spec.rb
@@ -1,10 +1,15 @@
 require 'spec_helper'
 
+DISABLE_SERVICE_NAME = 'service_name1 service_name_2'.freeze
+
 describe 'aws-parallelcluster-platform::disable_services' do
   for_all_oses do |platform, version|
     context "on #{platform}#{version}" do
       cached(:chef_run) do
-        runner(platform: platform, version: version).converge(described_recipe)
+        runner = ChefSpec::Runner.new do |node|
+          node.override['cluster']['disable_services'] = DISABLE_SERVICE_NAME
+        end
+        runner.converge(described_recipe)
       end
 
       it 'disables DLAMI multi eni helper' do
@@ -17,6 +22,14 @@ describe 'aws-parallelcluster-platform::disable_services' do
         is_expected.to disable_service('log4j-cve-2021-44228-hotpatch')
         is_expected.to stop_service('log4j-cve-2021-44228-hotpatch')
         is_expected.to mask_service('log4j-cve-2021-44228-hotpatch')
+      end
+
+      DISABLE_SERVICE_NAME.split().each do |service_name|
+        it "disables #{service_name}" do
+          is_expected.to disable_service(service_name)
+          is_expected.to stop_service(service_name)
+          is_expected.to mask_service(service_name)
+        end
       end
     end
   end